-bug: Deleting large number of hosts results in SQL errors due to MySQL buffer overrun -bug: Re-Apply Suggested Data Source names results in a SQL error for rrd_maximum='query_ifSpeed'

+++ cacti/branches/main/lib/api_graph.php
@@ -49,13 +49,24 @@ function api_graph_remove_multi($local_graph_ids) {
 			}
 
 			$i++;
+
+			if (($i % 1000) == 0) {
+				db_execute("DELETE FROM graph_templates_graph WHERE local_graph_id IN ($ids_to_delete)");
+				db_execute("DELETE FROM graph_templates_item WHERE local_graph_id IN ($ids_to_delete)");
+				db_execute("DELETE FROM graph_tree_items WHERE local_graph_id IN ($ids_to_delete)");
+				db_execute("DELETE FROM graph_local WHERE id IN ($ids_to_delete)");
+
+				$i = 0;
+				$ids_to_delete = "";
+			}
 		}
 
-		db_execute("DELETE FROM graph_templates_graph WHERE local_graph_id IN ($ids_to_delete)");
-		db_execute("DELETE FROM graph_templates_item WHERE local_graph_id IN ($ids_to_delete)");
-		db_execute("DELETE FROM graph_tree_items WHERE local_graph_id IN ($ids_to_delete)");
-		db_execute("DELETE FROM user_auth_perms WHERE item_id IN ($ids_to_delete) and type=" . PERM_GRAPHS);
-		db_execute("DELETE FROM graph_local WHERE id IN ($ids_to_delete)");
+		if ($i > 0) {
+			db_execute("DELETE FROM graph_templates_graph WHERE local_graph_id IN ($ids_to_delete)");
+			db_execute("DELETE FROM graph_templates_item WHERE local_graph_id IN ($ids_to_delete)");
+			db_execute("DELETE FROM graph_tree_items WHERE local_graph_id IN ($ids_to_delete)");
+			db_execute("DELETE FROM graph_local WHERE id IN ($ids_to_delete)");
+		}
 	}
 }
